+++ modules/features/discover/src/main/java/au/com/shiftyjelly/pocketcasts/discover/viewmodel/DiscoverViewModel.kt
@@ -127,6 +127,16 @@ class DiscoverViewModel @Inject constructor(
                 addPlaybackStateToList(it)
             }
     }
+    fun loadPodcasts(source: String, onPodcastsLoaded: (PodcastList) -> Unit) {
+        loadPodcastList(source).subscribeBy(
+            onNext = {
+                onPodcastsLoaded(it)
+            },
+            onError = {
+                Timber.e(it)
+            },
+        ).addTo(disposables)
+    }
 
     fun loadCarouselSponsoredPodcasts(
         sponsoredPodcastList: List<SponsoredPodcast>,