+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/to/Chapters.kt
@@ -3,6 +3,7 @@ package au.com.shiftyjelly.pocketcasts.models.to
 import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
+import kotlin.time.Duration
 
 data class Chapters(
     private val items: List<Chapter> = emptyList(),
@@ -20,8 +21,8 @@ data class Chapters(
     val lastChapter: Chapter?
         get() = items.getOrNull(items.size - 1)
 
-    fun getNextSelectedChapter(timeMs: Int): Chapter? {
-        val currentTimeFinal = if (timeMs < 0) 0 else timeMs
+    fun getNextSelectedChapter(time: Duration): Chapter? {
+        val currentTimeFinal = time.coerceAtLeast(Duration.ZERO)
         val items = if (FeatureFlag.isEnabled(Feature.DESELECT_CHAPTERS)) selectedItems else items
         for (chapter in items) {
             if (chapter.startTime > currentTimeFinal) {
@@ -31,7 +32,7 @@ data class Chapters(
         return null
     }
 
-    fun getPreviousSelectedChapter(timeMs: Int): Chapter? {
+    fun getPreviousSelectedChapter(time: Duration): Chapter? {
         if (items.isEmpty()) {
             return null
         }
@@ -39,12 +40,12 @@ data class Chapters(
         var lastChapter: Chapter? = null
         val items = if (FeatureFlag.isEnabled(Feature.DESELECT_CHAPTERS)) selectedItems else items
         for (chapter in items) {
-            if (chapter.containsTime(timeMs)) {
+            if (time in chapter) {
                 if (foundChapter != null) {
                     lastChapter = foundChapter
                 }
                 foundChapter = chapter
-            } else if (chapter.startTime <= timeMs) {
+            } else if (chapter.startTime <= time) {
                 lastChapter = chapter
             } else {
                 return lastChapter
@@ -53,56 +54,46 @@ data class Chapters(
         return lastChapter
     }
 
-    fun getChapter(time: Int): Chapter? {
-        if (isEmpty) {
-            return null
-        }
-        val finalTime = if (time < 0) 0 else time
-        var foundChapter: Chapter? = null
-        for (chapter in items) {
-            if (chapter.containsTime(finalTime)) {
-                foundChapter = chapter
-            }
-        }
-        return foundChapter
+    fun getChapter(time: Duration): Chapter? {
+        val finalTime = time.coerceAtLeast(Duration.ZERO)
+        return items.firstOrNull { chapter -> finalTime in chapter }
     }
 
-    fun getChapterIndex(time: Int): Int {
-        if (isEmpty) {
-            return -1
-        }
-        val finalTime = if (time < 0) 0 else time
-        var foundIndex = -1
-        var index = 0
-        for (chapter in items) {
-            if (chapter.containsTime(finalTime)) {
-                foundIndex = index
-            }
-            index++
-        }
-        return foundIndex
+    fun getChapterIndex(time: Duration): Int {
+        val finalTime = time.coerceAtLeast(Duration.ZERO)
+        return items.indexOfFirst { chapter -> finalTime in chapter }
     }
 
     fun getList(): List<Chapter> {
         return items
     }
 
-    fun getChapterSummary(time: Int): String {
+    fun getChapterSummary(time: Duration): String {
         val chapterSize = items.size
         val chapterIndex = getChapterIndex(time)
         return if (chapterIndex == -1) "" else "${chapterIndex + 1} of $chapterSize"
     }
 
-    fun isFirstChapter(time: Int): Boolean {
+    fun isFirstChapter(time: Duration): Boolean {
         return getChapterIndex(time) == 0
     }
 
-    fun isLastChapter(time: Int): Boolean {
+    fun isLastChapter(time: Duration): Boolean {
         return getChapterIndex(time) == items.size - 1
     }
 
-    fun updateDeselectedState(currentEpisode: BaseEpisode?) = this.copy(
-        items = this.getList().map { chapter ->
+    fun updateChaptersTimes(episodeDuration: Duration) = copy(
+        items = items.mapIndexed { index, chapter ->
+            when (index) {
+                0 -> chapter.copy(startTime = Duration.ZERO)
+                items.lastIndex -> chapter.copy(endTime = episodeDuration)
+                else -> chapter
+            }
+        },
+    )
+
+    fun updateDeselectedState(currentEpisode: BaseEpisode?) = copy(
+        items = items.map { chapter ->
             chapter.copy(selected = currentEpisode?.deselectedChapters?.contains(chapter.index) == false)
         },
     )