+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/PlaybackManager.kt
@@ -45,6 +45,7 @@ import au.com.shiftyjelly.pocketcasts.repositories.file.CloudFilesManager
 import au.com.shiftyjelly.pocketcasts.repositories.notification.NotificationHelper
 import au.com.shiftyjelly.pocketcasts.repositories.playback.LocalPlayer.Companion.VOLUME_DUCK
 import au.com.shiftyjelly.pocketcasts.repositories.playback.LocalPlayer.Companion.VOLUME_NORMAL
+import au.com.shiftyjelly.pocketcasts.repositories.podcast.ChapterManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PlaylistManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
@@ -87,10 +88,14 @@ import kotlin.coroutines.CoroutineContext
 import kotlin.math.abs
 import kotlin.math.min
 import kotlin.math.roundToInt
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.milliseconds
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.firstOrNull
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.reactive.asFlow
 import kotlinx.coroutines.rx2.asFlowable
@@ -125,6 +130,7 @@ open class PlaybackManager @Inject constructor(
     private val cloudFilesManager: CloudFilesManager,
     private val bookmarkManager: BookmarkManager,
     private val showNotesManager: ShowNotesManager,
+    private val chapterManager: ChapterManager,
     bookmarkFeature: BookmarkFeatureControl,
     private val playbackManagerNetworkWatcherFactory: PlaybackManagerNetworkWatcher.Factory,
     @ApplicationScope private val applicationScope: CoroutineScope,
@@ -796,6 +802,10 @@ open class PlaybackManager @Inject constructor(
         }
     }
 
+    private suspend fun seekToTimeMsInternal(duration: Duration) {
+        seekToTimeMsInternal(duration.inWholeMilliseconds.toInt())
+    }
+
     private suspend fun seekToTimeMsInternal(positionMs: Int) {
         LogBuffer.i(LogBuffer.TAG_PLAYBACK, "PlaybackService seekToTimeMsInternal %.3f ", positionMs.toDouble() / 1000.0)
         val episode = getCurrentEpisode()
@@ -893,7 +903,7 @@ open class PlaybackManager @Inject constructor(
         launch {
             val episode = getCurrentEpisode() ?: return@launch
             val currentTimeMs = getCurrentTimeMs(episode = episode)
-            playbackStateRelay.blockingFirst().chapters.getNextSelectedChapter(currentTimeMs)?.let { chapter ->
+            playbackStateRelay.blockingFirst().chapters.getNextSelectedChapter(currentTimeMs.milliseconds)?.let { chapter ->
                 seekToTimeMsInternal(chapter.startTime)
                 trackPlayback(AnalyticsEvent.PLAYBACK_CHAPTER_SKIPPED, SourceView.PLAYER)
             } ?: skipToEndOfLastChapter()
@@ -904,7 +914,7 @@ open class PlaybackManager @Inject constructor(
         launch {
             val episode = getCurrentEpisode() ?: return@launch
             val currentTimeMs = getCurrentTimeMs(episode)
-            playbackStateRelay.blockingFirst().chapters.getPreviousSelectedChapter(currentTimeMs)?.let { chapter ->
+            playbackStateRelay.blockingFirst().chapters.getPreviousSelectedChapter(currentTimeMs.milliseconds)?.let { chapter ->
                 seekToTimeMsInternal(chapter.startTime)
                 trackPlayback(AnalyticsEvent.PLAYBACK_CHAPTER_SKIPPED, SourceView.PLAYER)
             }
@@ -1519,23 +1529,41 @@ open class PlaybackManager @Inject constructor(
         updateCurrentPositionInDatabase()
     }
 
-    fun onMetadataAvailable(episodeMetadata: EpisodeFileMetadata) {
+    private fun onMetadataAvailable(episodeMetadata: EpisodeFileMetadata) {
         playbackStateRelay.blockingFirst().let { playbackState ->
-            val chapters = episodeMetadata.chapters
-            if (!chapters.isEmpty) {
-                chapters.getList().first().startTime = 0
-                chapters.getList().last().endTime = playbackState.durationMs
-            }
-
-            val chaptersWithDeselectState = chapters.updateDeselectedState(getCurrentEpisode())
+            val newChapters = episodeMetadata.chapters
+                .updateChaptersTimes(playbackState.durationMs.milliseconds)
+                .updateDeselectedState(getCurrentEpisode())
+            val chapters = maxOf(playbackState.chapters, newChapters) { a, b -> a.size.compareTo(b.size) }
 
             playbackStateRelay.accept(
                 playbackState.copy(
-                    chapters = chaptersWithDeselectState,
+                    chapters = chapters,
                     embeddedArtworkPath = episodeMetadata.embeddedArtworkPath,
                     lastChangeFrom = LastChangeFrom.OnMetadataAvailable.value,
                 ),
             )
+        }
+    }
+
+    @Volatile
+    private var observeChaptersJob: Job? = null
+
+    private fun onEpisodeChanged(episodeUuid: String) {
+        observeChaptersJob?.cancel()
+        observeChaptersJob = chapterManager.observerChaptersForEpisode(episodeUuid)
+            .onEach { onRemoteChaptersAvailable(it) }
+            .launchIn(this)
+    }
+
+    private fun onRemoteChaptersAvailable(remoteChapters: Chapters) {
+        playbackStateRelay.blockingFirst().let { playbackState ->
+            val newChapters = remoteChapters
+                .updateChaptersTimes(playbackState.durationMs.milliseconds)
+                .updateDeselectedState(getCurrentEpisode())
+            val chapters = maxOf(playbackState.chapters, newChapters) { a, b -> a.size.compareTo(b.size) }
+
+            playbackStateRelay.accept(playbackState.copy(chapters = chapters))
         }
     }
 
@@ -2106,6 +2134,7 @@ open class PlaybackManager @Inject constructor(
                 is PlayerEvent.MetadataAvailable -> onMetadataAvailable(event.metaData)
                 is PlayerEvent.PlayerError -> onPlayerError(event)
                 is PlayerEvent.RemoteMetadataNotMatched -> onRemoteMetaDataNotMatched(event.remoteEpisodeUuid)
+                is PlayerEvent.EpisodeChanged -> onEpisodeChanged(event.episodeUuid)
             }
         }
     }