+++ modules/services/sharedtest/src/main/java/au/com/shiftyjelly/pocketcasts/sharedtest/InMemoryFeatureFlagRule.kt
@@ -2,16 +2,13 @@ package au.com.shiftyjelly.pocketcasts.sharedtest
 
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.providers.InMemoryFeatureProvider
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.rules.TestWatcher
 import org.junit.runner.Description
 
-@ExperimentalCoroutinesApi
 class InMemoryFeatureFlagRule : TestWatcher() {
-
     override fun starting(description: Description) {
         FeatureFlag.initialize(
-            listOf(object : InMemoryFeatureProvider() {}),
+            listOf(InMemoryFeatureProvider()),
         )
     }